+++ src/compiler/typescript_to_rust/azle_generate_rearchitecture/src/traits/mod.rs
@@ -1,7 +1,11 @@
+pub use as_canister_method_annotation::AsCanisterMethodAnnotation;
+pub use as_exported_fn_decl::AsExportedFnDecl;
 pub use ident_value::IdentValue;
 pub use iter_with_peek::IterWithPeek;
 pub use to_ident::ToIdent;
 
+mod as_canister_method_annotation;
+mod as_exported_fn_decl;
 mod ident_value;
 mod iter_with_peek;
 mod to_ident;
